refactor(cli): split analyze tests into module file
